requirement.txt
sqlalchemy>=1.4
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
authlib==1.2.1
httpx==0.25.2
email-validator==2.1.0
python-dotenv==1.0.0